+++ src/components/feedback/__tests__/feedback-sent-rating.test.js
@@ -14,7 +14,13 @@ describe('Sent helpful rating - yes', () => {
         pathname: '/mapbox-gl-js/api/',
         hash: '#lnglat'
       }}
-      userName="decorah"
+      user={{
+        id: 'crocsfan19',
+        email: 'crocsfan19@mapbox.com',
+        plan: {
+          id: 'starter'
+        }
+      }}
       webhook={{
         production: '',
         staging: ''
@@ -26,7 +32,7 @@ describe('Sent helpful rating - yes', () => {
     expect(feedback.state()).toEqual({
       event: {
         event: 'Sent docs feedback',
-        userId: 'decorah',
+        userId: 'crocsfan19',
         properties: {
           environment: 'staging',
           helpful: true,
@@ -40,9 +46,10 @@ describe('Sent helpful rating - yes', () => {
             search: ''
           },
           page: { hash: '#lnglat', pathname: '/mapbox-gl-js/api/' },
+          plan: 'starter',
           section: undefined,
           site: 'dr-ui',
-          userId: 'decorah'
+          userId: 'crocsfan19'
         }
       },
       feedback: undefined,
@@ -60,7 +67,13 @@ describe('Sent helpful rating - no', () => {
         pathname: '/mapbox-gl-js/api/',
         hash: '#lnglat'
       }}
-      userName="decorah"
+      user={{
+        id: 'crocsfan19',
+        email: 'crocsfan19@mapbox.com',
+        plan: {
+          id: 'starter'
+        }
+      }}
       webhook={{
         production: '',
         staging: ''
@@ -72,7 +85,7 @@ describe('Sent helpful rating - no', () => {
     expect(feedback.state()).toEqual({
       event: {
         event: 'Sent docs feedback',
-        userId: 'decorah',
+        userId: 'crocsfan19',
         properties: {
           environment: 'staging',
           helpful: false,
@@ -86,9 +99,10 @@ describe('Sent helpful rating - no', () => {
             search: ''
           },
           page: { hash: '#lnglat', pathname: '/mapbox-gl-js/api/' },
+          plan: 'starter',
           section: undefined,
           site: 'dr-ui',
-          userId: 'decorah'
+          userId: 'crocsfan19'
         }
       },
       feedback: undefined,