+++ src/components/feedback/__tests__/feedback-test-cases.js
@@ -10,9 +10,9 @@ const testCases = {};
 
 testCases.basic = {
   component: Feedback,
-  description: 'Basic',
+  description: 'Basic (sends user information as `crocsfan19`)',
   props: {
-    site: 'Mapbox GL JS',
+    site: 'dr-ui',
     webhook: {
       staging:
         'https://evj5gwoa8j.execute-api.us-east-1.amazonaws.com/hookshot/webhook',
@@ -20,9 +20,16 @@ testCases.basic = {
         'https://2n40g6lyc9.execute-api.us-east-1.amazonaws.com/hookshot/webhook'
     },
     preferredLanguage: 'Swift',
+    user: {
+      id: 'crocsfan19',
+      email: 'crocsfan19@mapbox.com',
+      plan: {
+        id: 'starter'
+      }
+    },
     section: 'LngLat',
     location: {
-      pathname: '/mapbox-gl-js/api/',
+      pathname: '/dr-ui/feedback/',
       hash: '#lnglat'
     }
   }
@@ -30,10 +37,10 @@ testCases.basic = {
 
 testCases.type = {
   component: Feedback,
-  description: 'Change type',
+  description: 'Change type (sends anonymous user information)',
   props: {
     type: 'section',
-    site: 'Help',
+    site: 'dr-ui',
     webhook: {
       staging:
         'https://evj5gwoa8j.execute-api.us-east-1.amazonaws.com/hookshot/webhook',
@@ -46,9 +53,10 @@ testCases.type = {
 
 testCases.noSentry = {
   component: Feedback,
-  description: 'Does not send text feedback to Sentry',
+  description:
+    'Does not send text feedback to Sentry (sends anonymous user information)',
   props: {
-    site: 'Mapbox GL JS',
+    site: 'dr-ui',
     webhook: {
       staging:
         'https://evj5gwoa8j.execute-api.us-east-1.amazonaws.com/hookshot/webhook',
@@ -59,14 +67,14 @@ testCases.noSentry = {
     preferredLanguage: 'Swift',
     section: 'LngLat',
     location: {
-      pathname: '/mapbox-gl-js/api/',
+      pathname: '/dr-ui/api/',
       hash: '#lnglat'
     }
   }
 };
 
 testCases.common = {
-  description: 'Feedback placement',
+  description: 'Feedback placement (sends anonymous user information)',
   element: (
     <div>
       <TopbarSticker>
@@ -137,7 +145,7 @@ testCases.common = {
             </p>
 
             <Feedback
-              site="Help"
+              site="dr-ui"
               location={{}}
               webhook={{
                 staging: